airflow/requirements.txt
pyarrow
pendulum
psycopg2-binary
aiobotocore
asyncpg
astronomer-cosmos
dbt-snowflake
apache-airflow[amazon]
apache-airflow-providers-http
apache-airflow-providers-common-messaging
apache-airflow-providers-amazon
apache-airflow-providers-snowflake

# dbt-adapters==1.15.3
# dbt-common==1.25.0
# dbt-core==1.9.5
# dbt-extractor==0.6.0
# dbt-protos==1.0.317
# dbt-semantic-interfaces==0.7.4
# dbt-snowflake==1.9.4
# a2wsgi==1.10.8
# aenum==3.1.16
# aiobotocore==2.22.0
# aiohappyeyeballs==2.6.1
# aiohttp==3.12.4
# aioitertools==0.12.0
# aiologic==0.14.0
# aiosignal==1.3.2
# aiosqlite==0.21.0
# alembic==1.16.1
# annotated-types==0.7.0
# anyio==4.9.0
# apache-airflow==3.0.1
# apache-airflow-core==3.0.1
# apache-airflow-providers-amazon==9.8.0
# apache-airflow-providers-common-compat==1.7.0
# apache-airflow-providers-common-io==1.6.0
# apache-airflow-providers-common-messaging==1.0.2
# apache-airflow-providers-common-sql==1.27.1
# apache-airflow-providers-http==5.3.0
# apache-airflow-providers-smtp==2.1.0
# apache-airflow-providers-snowflake==5.8.0
# apache-airflow-providers-standard==1.2.0
# apache-airflow-task-sdk==1.0.1
# argcomplete==3.6.2
# asgiref==3.8.1
# asn1crypto==1.5.1
# astronomer-cosmos==1.10.1
# async-timeout==5.0.1
# asyncpg==0.30.0
# attrs==25.3.0
# backports-datetime-fromisoformat==2.0.3
# backports.strenum==1.3.1
# beautifulsoup4==4.13.4
# blinker==1.9.0
# boto3==1.37.3
# botocore==1.37.3
# cadwyn==5.3.3
# certifi==2025.4.26
# cffi==1.17.1
# charset-normalizer==3.4.2
# click==8.1.8
# cloudpickle==3.0.0
# colorlog==6.9.0
# cron-descriptor==1.4.5
# croniter==6.0.0
# cryptography==45.0.3
# Deprecated==1.2.18
# deprecation==2.1.0
# dill==0.4.0
# distlib==0.3.9
# dnspython==2.7.0
# email_validator==2.2.0
# exceptiongroup==1.3.0
# fastapi==0.115.12
# fastapi-cli==0.0.7
# filelock==3.18.0
# Flask==3.1.1
# frozenlist==1.6.0
# fsspec==2025.5.1
# googleapis-common-protos==1.70.0
# greenlet==3.2.2
# grpcio==1.71.0
# gunicorn==23.0.0
# h11==0.16.0
# httpcore==1.0.9
# httptools==0.6.4
# httpx==0.28.1
# idna==3.10
# importlib_metadata==8.6.1
# inflection==0.5.1
# isodate==0.7.2
# itsdangerous==2.2.0
# Jinja2==3.1.6
# jmespath==1.0.1
# jsonpath-ng==1.7.0
# jsonschema==4.23.0
# jsonschema-specifications==2025.4.1
# lazy-object-proxy==1.11.0
# libcst==1.7.0
# linkify-it-py==2.0.3
# lockfile==0.12.2
# lxml==5.3.2
# Mako==1.3.10
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# marshmallow==4.0.0
# mdurl==0.1.2
# methodtools==0.4.7
# more-itertools==10.7.0
# msgpack==1.1.0
# msgspec==0.19.0
# multidict==6.4.4
# numpy==1.26.4
# opentelemetry-api==1.33.1
# opentelemetry-exporter-otlp==1.33.1
# opentelemetry-exporter-otlp-proto-common==1.33.1
# opentelemetry-exporter-otlp-proto-grpc==1.33.1
# opentelemetry-exporter-otlp-proto-http==1.33.1
# opentelemetry-proto==1.33.1
# opentelemetry-sdk==1.33.1
# opentelemetry-semantic-conventions==0.54b1
# packaging==25.0
# pandas==2.1.4
# pathspec==0.12.1
# pendulum==3.1.0
# platformdirs==4.3.8
# pluggy==1.6.0
# ply==3.11
# propcache==0.3.1
# protobuf==5.29.4
# psutil==7.0.0
# psycopg2-binary==2.9.10
# pyarrow==20.0.0
# PyAthena==3.14.1
# pycparser==2.22
# pydantic==2.11.5
# pydantic_core==2.33.2
# Pygments==2.19.1
# PyJWT==2.10.1
# pyOpenSSL==25.1.0
# python-daemon==3.1.2
# python-dateutil==2.9.0.post0
# python-dotenv==1.1.0
# python-multipart==0.0.20
# python-slugify==8.0.4
# python3-saml==1.16.0
# pytz==2025.2
# PyYAML==6.0.2
# redshift-connector==2.1.7
# referencing==0.36.2
# requests==2.32.3
# requests-toolbelt==1.0.0
# retryhttp==1.3.3
# rich==14.0.0
# rich-argparse==1.7.1
# rich-toolkit==0.14.6
# rpds-py==0.25.1
# s3transfer==0.11.3
# sagemaker_studio==1.0.14
# scramp==1.4.5
# setproctitle==1.3.6
# shellingham==1.5.4
# six==1.17.0
# sniffio==1.3.1
# snowflake-connector-python==3.15.0
# snowflake-snowpark-python==1.32.0
# snowflake-sqlalchemy==1.7.3
# sortedcontainers==2.4.0
# soupsieve==2.7
# SQLAlchemy==1.4.54
# SQLAlchemy-JSONField==1.0.2
# SQLAlchemy-Utils==0.41.2
# sqlparse==0.5.3
# starlette==0.46.2
# structlog==25.3.0
# svcs==25.1.0
# tabulate==0.9.0
# tenacity==9.1.2
# termcolor==3.1.0
# text-unidecode==1.3
# tomli==2.2.1
# tomlkit==0.13.2
# typer==0.15.4
# types-requests==2.32.0.20250515
# typing-inspection==0.4.1
# typing_extensions==4.13.2
# tzdata==2025.2
# tzlocal==5.3.1
# uc-micro-py==1.0.3
# universal_pathlib==0.2.6
# urllib3==2.4.0
# uuid6==2024.7.10
# uvicorn==0.34.2
# uvloop==0.21.0
# virtualenv==20.31.2
# watchfiles==1.0.5
# watchtower==3.4.0
# websockets==15.0.1
# Werkzeug==3.1.3
# wirerope==1.0.0
# wrapt==1.17.2
# xmlsec==1.3.14
# yarl==1.20.0
# zipp==3.21.0
# 